requirements.txt
hydra-core==1.3.2 
hydra-colorlog==1.2.0 
hydra-optuna-sweeper==1.2.0

lightning==2.2.5 
lightning-utilities==0.11.3.post0 

wandb
rootutils 
rich
thop

transformers==4.42.0 
diffusers==0.29.0 
einops==0.7.0 
scipy==1.11.1 
spacy==3.7.1 
tensorboardX
git+https://github.com/openai/CLIP.git

# for mamba
causal-conv1d==1.3.0.post1

# for visualization
# imageio
# bpy
# matplotlib
# smplx
# h5py
# git+https://github.com/mattloper/chumpy
# imageio-ffmpeg
